# Implement equality for Values