Equality doesn't require the same amount of sophistication than ordering
comparison. Implement an Equal function that does just that but does it
much faster.

Benchmark results show up to 96% improvement:
```
benchmark                                                         old ns/op     new ns/op     delta
BenchmarkDeducedSimple-12                                         110083        103256        -6.20%
BenchmarkDeducedNested-12                                         374366        374160        -0.06%
BenchmarkDeducedNestedAcrossVersion-12                            392864        405525        +3.22%
BenchmarkLeafConflictAcrossVersion-12                             89112         89070         -0.05%
BenchmarkMultipleApplierRecursiveRealConversion-12                1564330       1574620       +0.66%
BenchmarkOperations/Pod/Create-12                                 103693        103970        +0.27%
BenchmarkOperations/Pod/Apply-12                                  291760        291317        -0.15%
BenchmarkOperations/Pod/Update-12                                 193419        190470        -1.52%
BenchmarkOperations/Pod/UpdateVersion-12                          261692        251966        -3.72%
BenchmarkOperations/Node/Create-12                                152047        155710        +2.41%
BenchmarkOperations/Node/Apply-12                                 499187        473901        -5.07%
BenchmarkOperations/Node/Update-12                                299271        279142        -6.73%
BenchmarkOperations/Node/UpdateVersion-12                         438723        403125        -8.11%
BenchmarkOperations/Endpoints/Create-12                           12246         11940         -2.50%
BenchmarkOperations/Endpoints/Apply-12                            915806        924080        +0.90%
BenchmarkOperations/Endpoints/Update-12                           7155675       285092        -96.02%
BenchmarkOperations/Endpoints/UpdateVersion-12                    14278150      544040        -96.19%
BenchmarkOperations/CustomResourceDefinition/Create-12            1312734       1288472       -1.85%
BenchmarkOperations/CustomResourceDefinition/Apply-12             3346591       3376864       +0.90%
BenchmarkOperations/CustomResourceDefinition/Update-12            10681243      1758764       -83.53%
BenchmarkOperations/CustomResourceDefinition/UpdateVersion-12     19069925      2202330       -88.45%
```